linspace TypeError correction

+++ PythonAPI/pycocotools/cocoevalLRP.py
@@ -437,7 +437,7 @@ class Params:
         self.imgIds = []
         self.catIds = []
         # np.arange causes trouble.  the data point on arange is slightly larger than the true value
-        self.confScores = np.linspace(.0, 1.00, np.round((1.00 - .0) / .01) + 1, endpoint=True)
+        self.confScores = np.linspace(.0, 1.00, int(np.round((1.00 - .0) / .01) + 1), endpoint=True)
         self.maxDets = 100
         self.areaRng = [[0 ** 2, 1e5 ** 2]]
         self.areaRngLbl = ['all']
